data model into the Reader
