+++ src/aqua/reader/reader.py
@@ -4,6 +4,7 @@ from contextlib import contextmanager
 import intake_esm
 import intake_xarray
 import xarray as xr
+from metpy.units import units
 
 from smmregrid import GridInspector
 
@@ -149,6 +150,8 @@ class Reader(FixerMixin, TimStatMixin):
         # store the kwargs for further usage
         self.kwargs = self._check_kwargs_parameters(kwargs, intake_vars)
 
+        # extend the unit registry
+        units_extra_definition()
         # Get fixes dictionary and find them
         self.fix = fix  # fix activation flag
         self.fixer_name = self.esmcat.metadata.get('fixer_name', None)
@@ -164,11 +167,11 @@ class Reader(FixerMixin, TimStatMixin):
         if self.fix:
             self.fixes_dictionary = load_multi_yaml(self.fixer_folder, loglevel=self.loglevel)
             self.fixes = self.find_fixes()  # find fixes for this model/exp/source
-            self.dst_datamodel = datamodel
+            self.tgt_datamodel = datamodel
             # Default destination datamodel
             # (unless specified in instantiating the Reader)
-            if not self.dst_datamodel:
-                self.dst_datamodel = self.fixes_dictionary["defaults"].get("dst_datamodel", None)
+            if not self.tgt_datamodel:
+                self.tgt_datamodel = self.fixes_dictionary["defaults"].get("dst_datamodel", None)
 
         # define grid names
         self.src_grid_name = self.esmcat.metadata.get('source_grid_name')
@@ -322,12 +325,13 @@ class Reader(FixerMixin, TimStatMixin):
                 fkind = "file from disk"
             data = log_history(data, f"Retrieved from {self.model}_{self.exp}_{self.source} using {fkind}")
 
-        if self.fix:
-            data = self.fixer(data, var)
 
         if not ffdb:  # FDB sources already have the index, already selected levels
             data = self._add_index(data)  # add helper index
             data = self._select_level(data, level=level)  # select levels (optional)
+
+        if self.fix:
+            data = self.fixer(data, var)
 
         # log an error if some variables have no units
         if isinstance(data, xr.Dataset) and self.fix:
@@ -1021,3 +1025,18 @@ class Reader(FixerMixin, TimStatMixin):
             for k, v in self.esmcat._request.items():
                 if k not in ["time", "param", "step", "expver"]:
                     print(f"  {k}: {v}")
+
+
+def units_extra_definition():
+    """Add units to the pint registry"""
+
+    # special units definition
+    # needed to work with metpy 1.4.0 see
+    # https://github.com/Unidata/MetPy/issues/2884
+    units._on_redefinition = 'ignore'
+    units.define('fraction = [] = Fraction = frac')
+    units.define('psu = 1e-3 frac')
+    units.define('PSU = 1e-3 frac')
+    units.define('Sv = 1e+6 m^3/s')
+    units.define("North = degrees_north = degreesN = degN")
+    units.define("East = degrees_east = degreesE = degE")
